random_joke_generator/python/requirements.txt
strands-agents>=0.1.0
flask>=2.3.0
flask-cors>=4.0.0
botocore[crt]>=1.29.0